Bulk ingest: single FOFN + Quickstart + Beta [VS-962] [VS-984] [VS-982] (#8404)
