+++ build.gradle
@@ -538,6 +538,10 @@ task collectBundleIntoDir(type: Copy) {
     from("scripts/sv", { into("scripts/sv") })
     from("scripts/cnv_wdl/", { into("scripts/cnv_wdl") })
     from("scripts/mutect2_wdl/", { into("scripts/mutect2_wdl") })
+    from("scripts/variantstore/") {
+        include 'wdl/extract/curate_input_array_files.py'
+        into("scripts/variantstore")
+    }
     from("scripts/dataproc-cluster-ui", { into("scripts/")})
     into "$buildDir/bundle-files-collected"
 }
