+++ scripts/variantstore/wdl/GvsAssignIds.wdl
@@ -128,8 +128,9 @@ task AssignIds {
   String bq_labels = "--label service:gvs --label team:variants --label managedby:assign_ids"
 
   command <<<
-    set -e
-    set -x
+    # Prepend date, time and pwd to xtrace log entries.
+    PS4='\D{+%F %T} \w $ '
+    set -o errexit -o nounset -o pipefail -o xtrace
 
     # make sure that sample names were actually passed, fail if empty
     num_samples=~{length(sample_names)}
