+++ scripts/variantstore/wdl/GvsCalculatePrecisionAndSensitivity.wdl
@@ -261,7 +261,7 @@ task Add_AS_MAX_VQS_SCORE_ToVcf {
     File input_vcf
     String output_basename
 
-    String docker = "us.gcr.io/broad-dsde-methods/variantstore:2023-06-23-alpine"
+    String docker = "us.gcr.io/broad-dsde-methods/variantstore:2023-07-14-alpine-a56ebf156"
     Int cpu = 1
     Int memory_mb = 3500
     Int disk_size_gb = ceil(2*size(input_vcf, "GiB")) + 50
