+++ scripts/variantstore/wdl/GvsQuickstartHailIntegration.wdl
@@ -14,6 +14,10 @@ workflow GvsQuickstartHailIntegration {
         String dataset_suffix = "hail"
         String? gatk_override
         String expected_output_prefix
+        String? sample_id_column_name ## Note that a column WILL exist that is the <entity>_id from the table name. However, some users will want to specify an alternate column for the sample_name during ingest
+        String? vcf_files_column_name
+        String? vcf_index_files_column_name
+        String? sample_set_name ## NOTE: currently we only allow the loading of one sample set at a time
     }
 
     String project_id = "gvs-internal"
@@ -28,6 +32,10 @@ workflow GvsQuickstartHailIntegration {
             gatk_override = gatk_override,
             interval_list = interval_list,
             expected_output_prefix = expected_output_prefix,
+            sample_id_column_name = sample_id_column_name,
+            vcf_files_column_name = vcf_files_column_name,
+            vcf_index_files_column_name = vcf_index_files_column_name,
+            sample_set_name = sample_set_name,
     }
 
     call ExtractAvroFilesForHail.GvsExtractAvroFilesForHail {
