+++ scripts/variantstore/wdl/extract/curate_bulk_import_tsv_test_files/input_samples_to_be_loaded_map.csv
@@ -1,0 +1,7 @@
+9,ERS4367804
+7,ERS4367801
+4,ERS4367798
+6,ERS4367800
+10,ERS4367805
+2,ERS4367796
+1,ERS4367795
